requirements.txt
flask == 2.0.1
flask-cors
flask-restx
flask-sqlalchemy
mysql-connector
flask-marshmallow
marshmallow >= 2.0.0
marshmallow-sqlalchemy